chore(CH): import changes from vamm-integration3
Imports the changes made in the original branch:
https://github.com/ComposableFi/composable/tree/escher-feature/vamm-integration3.
