+++ clearing-house/src/mock/mod.rs
@@ -1,14 +1,3 @@
-#[cfg(test)]
-pub mod runtime;
-
-#[cfg(test)]
-pub mod accounts;
-
-#[cfg(test)]
 pub mod assets;
-
-#[cfg(test)]
-pub mod vamm;
-
-#[cfg(test)]
-pub mod oracle;
+pub mod integration;
+pub mod unit;
