+++ clearing-house/src/tests/integration/mod.rs
@@ -1,2 +1,1 @@
-pub mod mock;
 pub mod tests;
