+++ clearing-house/src/tests/unit/instruments.rs
@@ -1,10 +1,10 @@
 use crate::{
-    mock::{
+    mock::assets::AssetId,
+    mock::unit::{
         self as mock,
-        assets::AssetId,
         runtime::{ExtBuilder, MarketId, Runtime, TestPallet, VammId},
     },
-    tests::{as_inner, bounded_decimal, zero_to_one_open_interval, Market, Position},
+    tests::unit::{as_inner, bounded_decimal, zero_to_one_open_interval, Market, Position},
     Direction,
 };
 use composable_traits::time::DurationSeconds;
