+++ clearing-house/src/tests/unit/settle_position.rs
@@ -1,9 +1,9 @@
 use crate::{
-    mock::{
+    mock::unit::{
         accounts::{ALICE, BOB},
         runtime::{ExtBuilder, Origin, Runtime, TestPallet, Vamm as VammPallet},
     },
-    tests::{
+    tests::unit::{
         as_balance, get_collateral, get_market, get_position, run_to_time,
         traders_in_one_market_context, with_market_context, with_trading_context,
     },
